abstract event subscription to separate class
